Created a new thread using inheritance from the thread class
